updating HOME View and normalizing styles between views
