+++ client/src/App.css
@@ -1,6 +1,6 @@
 /* Normalizing elements*/
 #content-container, #root, html, body {
-  height: 100%;
+  height: auto;
   width: 100%;
   padding: 0;
   margin: 0;
