+++ client/src/Views/HomeView.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function Home(props) {
+  return (
+    <React.Fragment>
+      <h2> It&apos;s Astra </h2>
+      <p> '*no login required*'</p>
+    </React.Fragment>
+  );
+}
+
+export default Home
