+++ client/src/Views/index.jsx
@@ -1,7 +1,9 @@
 import ListView from './ListView'
 import CreateView from './CreateView'
+import HomeView from './HomeView'
 
 export {
   ListView,
-  CreateView
+  CreateView,
+  HomeView
 }
